updates
merged:

  flora: 8112c4b2c507c770a37b8d1803f69a1e0067b87b
  faustoo: ed740b549e96cd221e2a4f0cd8620b5aa04d0ccb
  fusion809: 644933a9efef0562a3f2eebd70ce355f5aa4d2a4
  deadbeef: 53e0a41db6e7193bd83ef117f8055457f0990cb7
  gentoo-staging: 9292583023a6610905c79a4707db3ac6c63b3a31
  kit-fixups: 62e4f50b2e826665bff33c8f380b26c8f668aa86

+++ app-office/gnucash/gnucash-3.3.ebuild
@@ -1,22 +1,25 @@
+# Copyright 1999-2019 Gentoo Authors
 # Distributed under the terms of the GNU General Public License v2
 
 EAPI=6
 
 # google{test,mock} version
-PYTHON_COMPAT=( python3_{4,5,6} )
+GV="1.8.0"
+PYTHON_COMPAT=( python3_{5,6} )
 
-inherit cmake-utils gnome2-utils python-single-r1 xdg-utils
+inherit cmake-utils flag-o-matic gnome2-utils python-single-r1 xdg-utils
 
 DESCRIPTION="A personal finance manager"
 HOMEPAGE="http://www.gnucash.org/"
-SRC_URI="https://github.com/Gnucash/${PN}/releases/download/${PV}/${P}-1.tar.bz2"
+SRC_URI="https://github.com/Gnucash/${PN}/releases/download/${PV}/${P}.tar.bz2
+		 https://github.com/google/googletest/archive/release-${GV}.tar.gz -> gtest-${GV}.tar.gz"
 
 SLOT="0"
 LICENSE="GPL-2"
-KEYWORDS="~amd64 ~ppc ~ppc64 ~x86"
+KEYWORDS="amd64 ~ppc ~ppc64 ~x86"
 
 IUSE="aqbanking chipcard debug doc examples gnome-keyring mysql nls ofx postgres
-	  python quotes -register2 sqlite"
+	  python quotes -register2 sqlite +gui"
 REQUIRED_USE="
 	chipcard? ( aqbanking )
 	python? ( ${PYTHON_REQUIRED_USE} )"
@@ -27,20 +30,22 @@ REQUIRED_USE="
 RDEPEND="
 	>=dev-libs/glib-2.40.0:2
 	>=dev-libs/libxml2-2.7.0:2
-	>=dev-scheme/guile-2.0.11[regex]
 	>=sys-libs/zlib-1.1.4
-	>=x11-libs/gtk+-3.14.0:3
+	>=dev-scheme/guile-2.0.0:12=[regex]
 	dev-libs/boost:=[icu,nls]
 	dev-libs/icu:=
 	dev-libs/libxslt
-	gnome-base/dconf
-	net-libs/webkit-gtk:4=
 	aqbanking? (
 		>=net-libs/aqbanking-5[gtk,ofx?]
 		sys-libs/gwenhywfar[gtk]
 		chipcard? ( sys-libs/libchipcard )
 	)
 	gnome-keyring? ( >=app-crypt/libsecret-0.18 )
+	gui? (
+	  gnome-base/dconf
+	  net-libs/webkit-gtk:4=
+	  >=x11-libs/gtk+-3.14.0:3
+	)
 	mysql? (
 		dev-db/libdbi
 		dev-db/libdbi-drivers[mysql]
@@ -63,11 +68,10 @@ RDEPEND="
 "
 
 DEPEND="${RDEPEND}
-	>=dev-cpp/gtest-1.8.0
+	~dev-cpp/gtest-${GV}
 	>=sys-devel/gettext-0.19.6
 	dev-lang/perl
 	dev-perl/XML-Parser
-	gnome-base/gnome-common
 	sys-devel/libtool
 	virtual/pkgconfig
 "
@@ -77,16 +81,11 @@ PDEPEND="doc? (
 	gnome-extra/yelp
 )"
 
+PATCHES=( "${FILESDIR}"/${PN}-3.2-no-gui.patch )
+
 pkg_setup() {
 	use python && python-single-r1_pkg_setup
 	xdg_environment_reset
-}
-
-src_prepare() {
-	default
-	# avoid Werror.FL-5989.
-	sed -i -e 's/-Werror//' CMakeLists.txt || die "sed failed"
-	cmake-utils_src_prepare
 }
 
 src_configure() {
@@ -105,8 +104,11 @@ src_configure() {
 		-DWITH_OFX=$(usex ofx)
 		-DWITH_PYTHON=$(usex python)
 		-DWITH_SQL=${sql_on_off}
+		-DWITH_GNUCASH=$(usex gui)
 	)
 
+	append-cflags -Wno-error
+	append-cxxflags -Wno-error
 	cmake-utils_src_configure
 }
 
@@ -124,7 +126,10 @@ src_install() {
 	cmake-utils_src_install
 
 	rm "${ED%/}"/usr/share/doc/${PF}/README.dependencies || die
-	rm "${ED%/}"/usr/share/glib-2.0/schemas/gschemas.compiled || die
+
+	if use gui ; then
+	  rm "${ED%/}"/usr/share/glib-2.0/schemas/gschemas.compiled || die
+	fi
 
 	if use examples ; then
 		mv "${ED%/}"/usr/share/doc/gnucash \
@@ -143,8 +148,10 @@ src_install() {
 }
 
 pkg_postinst() {
-	gnome2_icon_cache_update
-	gnome2_schemas_update
+	if use gui ; then
+		gnome2_icon_cache_update
+		gnome2_schemas_update
+	fi
 	xdg_desktop_database_update
 	xdg_mimeinfo_database_update
 
@@ -155,8 +162,10 @@ pkg_postinst() {
 }
 
 pkg_postrm() {
-	gnome2_icon_cache_update
-	gnome2_schemas_update
+	if use gui ; then
+		gnome2_icon_cache_update
+		gnome2_schemas_update
+	fi
 	xdg_desktop_database_update
 	xdg_mimeinfo_database_update
 }
